Moved common hooks outside drop namespace.
